sockmap: minimal example of the sockmap BPF interface
The program accepts 3 connections and then splits the data from the
first one to the other two.  The data is split on a packet boundary and
the packets are alternated betwen the connections.  So for multiple
packets it looks like this:
 - packet 0 -> connection 1,
 - packet 1 -> connection 2,
 - packet 2 -> connection 1,
 - packet 3 -> connection 2, etc.

The easiest way to test it is using netcat:

[shell 0] $ src/sockmap 127.0.0.1 1024
[shell 1] $ nc 127.0.0.1 1024
[shell 2] $ nc 127.0.0.1 1024
[shell 3] $ nc 127.0.0.1 1024

Writing something on shell #1 will cause it to be redirected and output
to shell #2.  A second write will be redirectd it to shell #3, another
one to shell #2, then #3, etc.

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>