tools: test script for the NVMeoF/TCP parser
The script prepares various PDUs, fills them with random data and
verifies that the expected portions are recieved on the control
(headers, padding, digest) and data sockets.

It's intented to be used with the nvmf app that listens for three
connections and attached the necessary eBPF programs to split the
incoming data from the first socket to the other two.

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>